Move the proof-mutation variants into the innards namespace
Four of the mutation families lived in headers a user of the library includes,
two of them reachable from gcs/gcs.hh, so the public API advertised a way to
make the solver emit deliberately wrong proofs. They now sit in innards headers
in gcs::innards, which is what SubsetSumMutation and Am1FromPairsMutation always
did -- namespace-as-contract, the answer this tree already gave to #289, rather
than a pimpl that would put an indirection in every constraint object for a
naming problem.

Nothing about the coverage changes: in-tree tests include innards freely, and
the two Cumulative test files take the using-directive the rest of the tree
already has. `gcs::CumulativeProofMutation` now fails to compile, with the
compiler pointing at gcs::innards::CumulativeProofMutation, which is the whole
of the intended effect.

The issue says nothing reaches a mutation from an example binary or a command
line. That is no longer true: examples/rcpsp grew --mutate-makespan-bound in
#677, and the RCPSP bounds artefact uses it to require a refusal on the
instances the bound closes. It stays, and names the innards types at the use
site rather than hiding the crossing in a using-directive.

Compiling the hooks out of release builds stays rejected, and the reason is now
written down where someone would go looking for it rather than only in the
issue: the flags are lambda captures and cloned members, so a build without them
would ship a different closure from the one the tests exercised.

Closes #669
