+++ gcs/constraints/cumulative/cumulative_optional_test.cc
@@ -55,6 +55,7 @@ using fmt::println;
 #endif
 
 using namespace gcs;
+using namespace gcs::innards;
 using namespace gcs::test_innards;
 
 namespace
