+++ gcs/presolvers/cumulative_strengthening/cumulative_strengthening.hh
@@ -4,6 +4,7 @@
 #include <gcs/constraints/cumulative/cumulative.hh>
 #include <gcs/integer.hh>
 #include <gcs/presolver.hh>
+#include <gcs/presolvers/innards/cumulative_strengthening_mutations.hh>
 
 #include <cstddef>
 #include <memory>
@@ -100,59 +101,6 @@ namespace gcs
     };
 
     /**
-     * \brief Deliberate corruptions of the strengthening derivation, for
-     * testing only. VeriPB must reject each of them.
-     *
-     * \ingroup Presolvers
-     */
-    namespace cumulative_strengthening_mutation
-    {
-        /// Emit the honest derivation.
-        struct None
-        {
-        };
-
-        /// Claim a capacity one below the largest load the heights can reach.
-        /// The "bound + 1 must fail" check for this rule: it corrupts the
-        /// conclusion rather than the route to it.
-        struct ClaimOneBetter
-        {
-        };
-
-        /// Take the divisibility fast path with a divisor that does not divide
-        /// every height. Dividing is a sound proof step whatever the divisor,
-        /// so this produces a perfectly valid line that simply is not the one
-        /// the derived constraint was told it had --- which nothing catches
-        /// except the `ia` step pinning each row's content.
-        struct BogusDivisor
-        {
-        };
-
-        /// Raise the tallest task that did *not* qualify for it. The pairwise
-        /// conflict test is the only thing standing between the rule and an
-        /// unsound constraint, and this is what says so: the derivation itself
-        /// runs honestly and every step of it is sound, on a set that is wrong.
-        /// Needs a fixture with a task that fails the test, which is what an
-        /// R1 control fixture is.
-        struct RaiseUnentitled
-        {
-        };
-
-        /// Take one more than the largest step the division survives, on the
-        /// first step of each raise. The step lands on a sound but weaker line,
-        /// every later step compounds it, and the row's own `ia` pin is what
-        /// rejects. Needs a raise with a step to spare, and throws rather than
-        /// passing quietly if given one that has none.
-        struct RaiseTooFast
-        {
-        };
-    }
-
-    using CumulativeStrengtheningMutation = std::variant<cumulative_strengthening_mutation::None, cumulative_strengthening_mutation::ClaimOneBetter,
-        cumulative_strengthening_mutation::BogusDivisor, cumulative_strengthening_mutation::RaiseUnentitled,
-        cumulative_strengthening_mutation::RaiseTooFast>;
-
-    /**
      * \brief Strengthen each posted Cumulative by integrality, posting the
      * strengthened version as a *derived* constraint whose per-time capacity
      * rows are proved from the donor's.
@@ -211,7 +159,7 @@ namespace gcs
         long long _max_dynamic_programming_states;
         long long _max_raise_lines;
         CumulativeRules _rules;
-        CumulativeStrengtheningMutation _mutation;
+        innards::CumulativeStrengtheningMutation _mutation;
 
     public:
         /**
@@ -274,7 +222,7 @@ namespace gcs
 
         /// Corrupt one step of the emitted derivation. For tests only, which
         /// assert that VeriPB rejects the result.
-        auto with_proof_mutation(CumulativeStrengtheningMutation mutation) -> CumulativeStrengthening &;
+        auto with_proof_mutation(innards::CumulativeStrengtheningMutation mutation) -> CumulativeStrengthening &;
 
         [[nodiscard]] virtual auto run(Problem &, innards::Propagators &, innards::State &, innards::ProofLogger * const) -> bool override;
         [[nodiscard]] virtual auto clone() const -> std::unique_ptr<Presolver> override;
